+++ src/main/java/com/planetrush/planetrush/core/config/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package com.planetrush.planetrush.core.config.querydsl;
+package com.planetrush.planetrush.core.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
