+++ src/main/java/com/planetrush/planetrush/core/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.planetrush.planetrush.core.config.s3;
+package com.planetrush.planetrush.core.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
