+++ src/main/java/com/planetrush/planetrush/core/config/SwaggerConfig.java
@@ -1,11 +1,10 @@
-package com.planetrush.planetrush.core.config.swagger;
+package com.planetrush.planetrush.core.config;
 
 import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.info.Contact;
 import io.swagger.v3.oas.annotations.info.Info;
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
@@ -16,9 +15,7 @@ import io.swagger.v3.oas.models.security.SecurityScheme;
 	info = @Info(
 		title = "플래닛러쉬 API",
 		description = "<h3>플래닛러쉬 API</h3>",
-		version = "v1",
-		contact = @Contact(
-			name = "김도은(팀장), 김경민, 박정의, 심종한, 안홍찬, 이강현")
+		version = "v1"
 	)
 )
 @Configuration
